refactor: move github package from pkg/ to internal/
